# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.peerUpdateBody:
    properties:
      allowed_ip:
        type: string
      buy_date:
        type: string
      config_end_point:
        type: string
      data_limit:
        type: number
      email:
        type: string
      expire_date:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
  models.Peer:
    properties:
      allowed_ip:
        type: string
      buy_date:
        type: string
      config_end_point:
        type: string
      data_limit:
        type: number
      email:
        type: string
      end_point:
        type: string
      expire_date:
        type: string
      is_active:
        type: boolean
      last_handshake:
        type: string
      name:
        type: string
      phone:
        type: string
      usage:
        type: number
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a Wireguard API server.
  title: Wireguard API
  version: "1.0"
paths:
...
  /peers/{name}:
    delete:
      description: Delete a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer deleted
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Peer not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a peer
      tags:
      - Peers
    get:
      description: Retrieve information about a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer information
          schema:
            $ref: '#/definitions/models.Peer'
        "404":
          description: Peer not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get information about a peer
      tags:
      - Peers
    put:
      consumes:
      - application/json
      description: Update a peer's information by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      - description: Updated peer information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.peerUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Peer updated
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Fields cannot be empty
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a peer's information
      tags:
      - Peers
  /peers/{name}/conf:
    get:
      description: Retrieve a peer's config by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Peer config
          schema:
            type: string
        "404":
          description: Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a peer's config
      tags:
      - Peers
  /peers/{name}/pause:
    put:
      description: Pause a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer paused
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer already paused" "Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Pause a peer
      tags:
      - Peers
  /peers/{name}/qr:
    get:
      description: Retrieve a peer's QR code by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: Peer QR code
          schema:
            type: string
        "404":
          description: Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a peer's QR code
      tags:
      - Peers
  /peers/{name}/reset:
    put:
      description: Reset a peer's usage by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer usage reset
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reset a peer's usage
      tags:
      - Peers
  /peers/{name}/resume:
    put:
      description: Resume a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer resumed
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer already active" "Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Resume a peer
      tags:
      - Peers
...
  /test/peers/{name}/pause:
    put:
      description: Pause a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer paused
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer already paused" "Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Pause a peer
      tags:
      - Peers
      - Test
  /test/peers/{name}/reset:
    put:
      description: Reset a peer's usage by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer usage reset
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reset a peer's usage
      tags:
      - Peers
      - Test
  /test/peers/{name}/resume:
    put:
      description: Resume a peer by name.
      parameters:
      - description: Peer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Peer resumed
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Peer already active" "Peer not found" "System not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Resume a peer
      tags:
      - Peers
      - Test
...
swagger: "2.0"